@@ -12,17 +12,14 @@
     }
   ],
   "require": {
-    "php": "^8.0",
+    "php": ">=8.1",
     "ext-bcmath": "*"
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "^3.13",
-    "phpstan/phpstan": "^1.9",
-    "phpstan/phpstan-phpunit": "^1.3",
-    "phpstan/phpstan-webmozart-assert": "^1.2",
-    "phpunit/phpunit": "^9.5",
+    "friendsofphp/php-cs-fixer": "^3.51",
+    "phpunit/phpunit": "^10.5 | ^11.0",
     "psalm/plugin-phpunit": "^0.18",
-    "squizlabs/php_codesniffer": "^3.7",
+    "squizlabs/php_codesniffer": "^3.9",
     "vimeo/psalm": "^5.4"
   },
   "autoload": {
@@ -41,7 +38,6 @@
     "cs-fix": "php-cs-fixer --using-cache=no fix",
     "test": "export XDEBUG_MODE=coverage && vendor/bin/phpunit --configuration phpunit.xml",
     "psalm": "vendor/bin/psalm --no-cache",
-    "analyze": "vendor/bin/phpstan analyse --xdebug --configuration phpstan.neon.dist",
     "check": [
       "@phpcs",
       "@cs-check",
